vibrator: add support to use led-trigger interface
The existing vibrator HAL uses the timed_output interface which will
never be moved out of kernel's staging tree. The kernel already has a
similar feature with GPIO controlled LEDs and LED triggers. While a
vibrator is not an LED, the control model is the same and can be reused.
Additionally, DT bindings already exist for GPIO controlled LEDs.

This adds support for the LED device in addition to the timed_output
device. It uses the "transient" trigger for the LED device named
"vibrator".

Change-Id: I90d52d6e50d83d255f484db148064069504f78e0
Signed-off-by: Rob Herring <robh@kernel.org>